Migrated NewGame to it's own file
