+++ core/consts.ts
@@ -1,0 +1,3 @@
+enum worldStats {
+  
+}
