+++ core/utils.ts
@@ -1,0 +1,3 @@
+export function rollDice () {
+  return Math.floor((Math.random() * 6) + 1)
+}
